requirements.txt
python_qbittorrent>=0.4.3
python-dotenv>=1.0.0
humanfriendly>=10.0
